Add CLI to estimate EMOS coefficients (#859)
* Current status after reworking the inputs and outputs for the EstimateCoefficientsForEnsembleCalibration and ApplyCoefficientsFromEnsembleCalibration classes, so that a coefficients cube is output when the coefficients are estimated and this cube is used when applying the coefficients. Unit tests have been updated.

* Initial attempt at CLI for estimating EMOS coefficients.

* Add the creation of a coefficient cube, as part of the estimation of the coefficients.

* Add docstrings to unit tests.

* Remove handling of multiple timesteps from ensemble calibration, so that the unit tests are simplified and so as to simplify further changes.

* Current status of ensemble calibration work.

* Modifications, for example to make use of the current cycletime as an input, rather than using a current forecast cube.

* Remove current prototype CLI for estimating the EMOS coefficients for addition in a later branch.

* First addition of a estimate EMOS coefficients CLI.

* Addition of tests for estimate-emos-coefficients CLI.

* Edits to incorporate the estimate-emos-coefficients CLI.

* Amend cycletime docstring.

* Minor amendments following merge.

* Minor edit to unit tests following enforcing coefficients to be float32.

* Update 01-help.bats.

* Edit to specify the datatype for the coefficient_index.

* Correction to 03-trunc_gaussian.bats.

* Second review changes.

+++ lib/improver/ensemble_calibration/ensemble_calibration.py
@@ -197,7 +197,7 @@ class ContinuousRankedProbabilityScoreMinimisers(object):
                        self.MAX_ITERATIONS, optimised_coeffs.message))
             warnings.warn(msg)
         calculate_percentage_change_in_last_iteration(optimised_coeffs.allvecs)
-        return optimised_coeffs.x
+        return optimised_coeffs.x.astype(np.float32)
 
     def normal_crps_minimiser(
             self, initial_guess, forecast_predictor, truth, forecast_var,
@@ -449,7 +449,7 @@ class EstimateCoefficientsForEnsembleCalibration(object):
             raise ValueError(msg)
 
         coefficient_index = iris.coords.DimCoord(
-            np.arange(len(optimised_coeffs)),
+            np.arange(len(optimised_coeffs), dtype=np.int32),
             long_name="coefficient_index", units="1")
         coefficient_name = iris.coords.AuxCoord(
             coeff_names, long_name="coefficient_name", units="no_unit")
